+++ src/detectorist/detectorist_app.py
@@ -989,6 +989,7 @@ class DetectoristApp(QMainWindow):
             menu.addAction(self.ui.crop_and_export_selected_images_action)
             menu.addAction(self.ui.copy_export_remove_action)
             menu.addAction(self.ui.remove_selected_images_from_list_action)
+            menu.addAction(self.ui.clear_image_list_action)
             menu.addSeparator()
             menu.addAction(self.ui.copy_filenames_to_clipboard_action)