+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [0.10.0] - 2026-07-08
 ### Added
 - The welcome screen now lists recently opened folders to provide a quick way to reopen them.
 - New "Actions > Open output folder" menu item reveals the most recent export directory.